+++ suricatalog/canned.py
@@ -9,7 +9,7 @@ import locale
 
 from textual.app import App
 
-from suricatalog.flow_apps import FlowApp
+from suricatalog.flow_app import FlowApp
 from suricatalog.hostdatause_app import HostDataUse
 from suricatalog.topuser_app import TopUserApp
 from suricatalog.oneshot_app import OneShotApp
